+++ src/lib.rs
@@ -29,7 +29,7 @@ mod tests {
     use super::*;
     use mio::net::{TcpStream,TcpListener};
     use std::io::{Write,Read};
-    use std::time::Duration;
+    use std::time::{Duration,Instant};
     use std::io;
     use std::str;
     use native_tls::{TlsConnector};
@@ -38,24 +38,33 @@ mod tests {
     struct Param {
         chosen: Option<String>,
         is_https: bool,
+        proxy_client: bool,
         http_hosts: Vec<String>,
         https_hosts: Vec<String>,
     }
 
+    #[derive(PartialEq)]
+    enum Resp<'a> {
+        Done,
+        Bytes(&'a[u8])
+    }
+
     // Receive list of hosts.
     // Return slices.
-    fn get_http(mut fiber: Fiber<Param,&[u8]>, p: Param) -> Option<&[u8]> {
+    fn get_http(mut fiber: Fiber<Param,Resp>, p: Param) -> Option<Resp> {
         // We will read in 500B chunks
-        let mut v = [0u8;500];
+        let mut v = [0u8;2000];
         let host = p.chosen.unwrap();
+
+        println!("get_http {}", host);
 
         let timeout = if p.is_https {
             let connector = TlsConnector::builder().unwrap().build().unwrap();
             fiber.tcp_tls_connect(connector, host.as_str());
             // https requires longer timeout
-            fiber.socket_timeout(Some(Duration::from_millis(1000)));
+            fiber.socket_timeout(Some(Duration::from_millis(2000)));
         } else {
-            fiber.socket_timeout(Some(Duration::from_millis(500)));
+            fiber.socket_timeout(Some(Duration::from_millis(2000)));
         };
 
         // We want to time out so use keep-alive
@@ -67,31 +76,33 @@ mod tests {
             match fiber.read(&mut v[..]) {
                 Ok(sz) => {
                     // Return slice to parent, directly from our stack!
-                    fiber.resp_chunk(&v[0..sz]);
+                    fiber.resp_chunk(Resp::Bytes(&v[0..sz]));
                 }
                 Err(e) => {
-                    assert_eq!(e.kind(), io::ErrorKind::TimedOut);
+                    // assert_eq!(e.kind(), io::ErrorKind::TimedOut);
                     break;
                 }
             }
         }
-        println!("Client fiber closing");
-        None
+        println!("Client fiber closing {}", p.proxy_client);
+        Some(Resp::Done)
     }
 
-    fn rand_http_proxy(mut fiber: Fiber<Param,&[u8]>, p: Param) -> Option<&[u8]> {
+    fn rand_http_proxy(mut fiber: Fiber<Param,Resp>, p: Param) -> Option<Resp> {
         fiber.socket_timeout(Some(Duration::from_millis(500)));
 
         // Pick a random host from our list.
         let chosen = rand::random::<usize>() % p.http_hosts.len();
         // Pick http or https.
-        let port = if rand::random::<u8>() % 2 == 0 { 80 } else { 443 };
+        //let port = if rand::random::<u8>() % 2 == 0 { 80 } else { 443 };
+        let port = 80;
         let p1 = if port == 443 {
             Param {
                 chosen: Some(p.https_hosts[chosen].clone()),
                 is_https: port == 443,
                 http_hosts: Vec::new(),
                 https_hosts: Vec::new(),
+                proxy_client: true,
             }
         } else {
              Param {
@@ -99,6 +110,7 @@ mod tests {
                 is_https: port == 443,
                 http_hosts: Vec::new(),
                 https_hosts: Vec::new(),
+                proxy_client: true,
             }
         };
         let addr = if let &Some(ref ch) = &p1.chosen {
@@ -106,24 +118,31 @@ mod tests {
         } else {
             panic!("")
         };
-        println!("Returning: {}{}", if port == 443 { "https://" } else { "http://" },  addr);
+        let a = Instant::now();
         // Start connection to host
-        let client_sock = TcpStream::from_stream(::std::net::TcpStream::connect(addr).unwrap()).unwrap();
+        let client_sock = TcpStream::from_stream(::std::net::TcpStream::connect(addr.clone()).unwrap()).unwrap();
         // Join our fiber to it. This way we can receive its output.
         fiber.join_tcp(client_sock, get_http, p1);
+        println!("Returning ({}ms,{}ms): {}{}", 
+            a.elapsed().subsec_nanos() / 1000000, 
+            b.elapsed().subsec_nanos() / 1000000,
+            if port == 443 { "https://" } else { "http://" },  addr);
 
         // Fibers can stream response to parent. So we iterate on responses.
         // We could also create multiple children and iterate on all of them.
-        while let Some(slice) = fiber.get_child() {
-            fiber.write(slice);
+        while let Some(resp) = fiber.get_child() {
+            if let Resp::Bytes(slice) = resp {
+                println!("Server got {}", slice.len());
+                fiber.write(slice);
+            }
         }
         println!("Server socket fiber closing");
         // return empty slice, so main stack knows a server connection has closed
-        Some(&[])
+        None
     }
 
     // Accept sockets in an endless loop.
-    fn sock_acceptor(mut fiber: Fiber<Param,&[u8]>, p: Param) -> Option<&[u8]> {
+    fn sock_acceptor(mut fiber: Fiber<Param,Resp>, p: Param) -> Option<Resp> {
         loop {
             // If no sockets available, fiber will be scheduled out for execution until something connects. 
             match fiber.accept_tcp() {
@@ -148,6 +167,7 @@ mod tests {
         let p = Param {
             chosen: None,
             is_https: false,
+            proxy_client: false,
             http_hosts: vec!["www.liquiddota.com".to_string(),"www.google.com".to_string(),
                 "www.sqlite.org".to_string(),"edition.cnn.com".to_string()],
             https_hosts: vec!["www.reddit.com".to_string(), "www.google.com".to_string(),
@@ -155,18 +175,33 @@ mod tests {
         };
         // Start our fiber poller.
         // Set this stack lower to see some SIGBUS action.
-        let poll:Poller<Param,&[u8]> = Poller::new(Some(4096*10)).unwrap();
+        let poll:Poller<Param,Resp> = Poller::new(Some(4096*10)).unwrap();
         // Start a TCP listener socket
         let listener = TcpListener::bind(&"127.0.0.1:10000".parse().unwrap()).unwrap();
         // Create a fiber from it. Listener socket will use sock_acceptor function.
         poll.new_listener(listener, sock_acceptor, p).unwrap();
         // Poll for 3 requests before exiting.
-        let mut reqs_remain = 3;
+        let mut reqs_remain = 20;
+        for i in 0..reqs_remain {
+            let p = Param {
+                chosen: Some("127.0.0.1:10000".to_string()),
+                is_https: false,
+                proxy_client: false,
+                http_hosts: Vec::new(),
+                https_hosts: Vec::new(),
+            };
+            let client_sock = TcpStream::from_stream(::std::net::TcpStream::connect("127.0.0.1:10000").unwrap()).unwrap();
+            poll.new_tcp(client_sock, get_http, p);
+        }
         while reqs_remain > 0 {
             if poll.poll(Duration::from_millis(10)) {
                 while let Some(r) = poll.get_response() {
-                    println!("Finished executing, req_remain: {}", reqs_remain);
-                    reqs_remain -= 1;
+                    if Resp::Done == r {
+                        reqs_remain -= 1;
+                        println!("Finished executing, req_remain: {}", reqs_remain);
+                    } else if let Resp::Bytes(slice) = r {
+                        println!("Main stack got {} bytes", slice.len());
+                    }
                 }
 
                 // we arent using